main: move InterruptHandler to util
License: MIT
